requirements.txt
gradio>=4.0.0
librosa>=0.10.1
openai==1.55.3
httpx==0.27.2
replicate==0.22.0
pydub>=0.25.1
numpy>=1.24.0
matplotlib>=3.7.0
mediapipe>=0.10.8
opencv-python-headless>=4.8.0
python-dotenv
pygame
requests
opencv-python